+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 - [Optimize: remove discovery module useless code](https://github.com/Tencent/spring-cloud-tencent/pull/597)
 - [Fix issue 579:Report the labels in request when report the result of invocation by RestTemplate](https://github.com/Tencent/spring-cloud-tencent/pull/600)
+- [Optimize: optimize configuration conditional & optimize config data tips"](https://github.com/Tencent/spring-cloud-tencent/pull/605)